Move contents of Color+Button.swift into Color.swift
