--- Sources/DNA/Color/Color+Button.swift
@@ -1,27 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import Foundation
-
-extension UIColor {
-    public class var callToActionButtonHighlightedBodyColor: UIColor {
-        return UIColor(r: 0, g: 79, b: 201)!
-    }
-
-    public class var destructiveButtonHighlightedBodyColor: UIColor {
-        return UIColor(r: 201, g: 79, b: 0)!
-    }
-
-    public class var defaultButtonHighlightedBodyColor: UIColor {
-        return UIColor(r: 241, g: 249, b: 255)!
-    }
-
-    public class var linkButtonHighlightedTextColor: UIColor {
-        return UIColor(r: 0, g: 79, b: 201)!
-    }
-
-    public class var flatButtonHighlightedTextColor: UIColor {
-        return UIColor(r: 0, g: 79, b: 201)!
-    }
-}
