+++ Sources/DNA/Color/Color.swift
@@ -65,3 +65,27 @@ extension UIColor {
         self.init(red: r / 255.0, green: g / 255.0, blue: b / 255.0, alpha: a)
     }
 }
+
+// MARK: - Button
+
+extension UIColor {
+    public class var callToActionButtonHighlightedBodyColor: UIColor {
+        return UIColor(r: 0, g: 79, b: 201)!
+    }
+
+    public class var destructiveButtonHighlightedBodyColor: UIColor {
+        return UIColor(r: 201, g: 79, b: 0)!
+    }
+
+    public class var defaultButtonHighlightedBodyColor: UIColor {
+        return UIColor(r: 241, g: 249, b: 255)!
+    }
+
+    public class var linkButtonHighlightedTextColor: UIColor {
+        return UIColor(r: 0, g: 79, b: 201)!
+    }
+
+    public class var flatButtonHighlightedTextColor: UIColor {
+        return UIColor(r: 0, g: 79, b: 201)!
+    }
+}
